add setting to view other album photos of people The intent is to let you select a person or multiple persons, like you currently can. And if this new feature is enabled, then the photos returned will be from any album that any of those tagged people are in.
For example, I have Bob and Jane added in the People list. Without this new feature,
it will show any photo where Bob OR Jane are tagged.
With the new feature, it will find any album that Bob OR Jane are tagged in, and
show any image from that album.

I want to have people come over, and I add that person as one of the tagged people.
But I don't want to limit photos to just the ones that show their face. I want to include
other photos from any events that person was in (ie an album).

As long as they are tagged in at least one photo from an album, all photos from that
album are fair game to be shown. If they aren't tagged in any photo from an album,
then none of the photos from that album are shown.